Allow installing the user's registered public SSH keys
One quite valid concern with running tmate in a GitHub workflow is that
_anybody_ can connect to the opened session. This is particularly
problematic when secret information was laid down to disk.

To address this concern, this commit introduces the Boolean input
`limit-access-to-actor`. When it is set to `true`, the Action will
install the public SSH key(s) registered with the GitHub profile of the
actor (i.e. the user triggering the workflow).

For more information how publich SSH keys can be registered, see:
https://docs.github.com/en/github/.authenticating-to-github/adding-a-new-ssh-key-to-your-github-account

If no SSH keys are registered, and the `limit-access-to-actor` input is
set to `true`, the Action will complain loudly and fail.

This closes https://github.com/mxschmitt/action-tmate/issues/2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>