Install @actions/github and @octokit/rest
We will need them in the next commit, where we will implement a new
option to allow installing the public SSH key(s) that the user has
registered with their GitHub profile.

Note: since we want to run the tests in a GitHub workflow, which sets
`GITHUB_EVENT_PATH`, the constructor of @actions/github would want to
initialize the environment as if running in a workflow (which it is),
but we do not want that (because this is a CI test run). Therefore, we
need to make sure that it is unset in the `test` script in
`package.json`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>