Make sure to start `bash.exe` in the workspace
The MSYS2 project provides a feature (mysteriously called
`CHERE_INVOKING`) to ask the login shell _not_ to change working
directory to the user's home directory.

However, as soon as the login shell saw that it was not supposed to
change the working directory, it _unsets_ `CHERE_INVOKING` so that
subsequently-called login shells _do_ switch to the home directory.

This interferes with what we are trying to do here: the first login
shell will be spawned by `execShellCommand()`. _Within_ that login
shell, we call `tmate new-session`, which will open a TMate session that
opens a new TMate window, within which it will run _another_ login
shell.

To let that "inner login shell" know that we _still_ do not want the
working directory to be changed, let's make sure that `tmate` is run
with `CHERE_INVOKING` set to a non-empty value.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>