fix: turn `errexit` off in the shell sessions (#64)
* DRY up the `tmate` invocations

We always pass the `-S /tmp/tmate.sock` option; Let's not repeat this
(and risk forgetting it anywhere, or risk typos, or... something).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* Work around Actions adding `set -e` to the Bash profile

When `set -e` (or more verbosely, `set -o errexit`) is in effect, any
failing command will automatically exit the shell. This is quite
convenient for scripts.

It is quite awful, however, for interactive sessions: even a tiny typo
will send the session to the happy hunting ground.

In fact, it is actually not _all_ that good an option to use, anyway,
see https://mywiki.wooledge.org/BashFAQ/105 for more details.

Therefore, no `/etc/profile` will set this, as it would render the
interactive sessions useless.

However, `conda-incubator/setup-miniconda` does this. Granted, the
assumption that GitHub workflows do not run interactively is an easy one
to make. And it was correct... until `action-tmate` came along.

Let's work around this by _specifically_ turning off `errexit`.

Since we do not want to add hackery on top of hackery, we will not edit
`~/.profile` but instead write our own rcfile and pass it onto the
default command `tmate` runs.

This fixes https://github.com/mxschmitt/action-tmate/issues/37.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>