Fix Jekyll formatting of code block
As I noted in [this][1] comment on #78, the formatting on the [Jekyll generated page][2] is wrong as
(I think) it's interpreting the `${{ ... }}` as a Liquid tag.

We need to wrap the code block in a [`{% raw %} .. {% endraw %}`][3], but the only way I could think
of doing this without it being visible in the GitHub markdown was to enclose it in a standard HTML
comment. Hacky, but seems to work!

[1]: https://github.com/mxschmitt/action-tmate/pull/78#issuecomment-821243039
[2]: https://mxschmitt.github.io/action-tmate/
[3]: https://blog.slaks.net/2013-06-10/jekyll-endraw-in-code/