Fix action-tmate on Windows (#151)
windows: fix `tmate` install procedure

The `tmate` package in MSYS2 actually depends on other packages, too,
e.g. `libopenssl`. These dependencies are automatically installed unless
they are installed already.

Crucially, though, these dependencies are not _upgraded_ automatically.

This became a problem when MSYS2 switched from OpenSSL v1.1.x to v3.x
because the version number is sadly manifested in the DLL's file name,
meaning that running `tmate.exe` would fail e.g. if it wants to find an
`msys-crypto-3.dll` (but only `msys-crypto-1.1.dll` is there).

The obvious solution would be to pass the `-u` flag to the package
manager, to upgrade the already-installed packages (as in `pacman -Syu`,
as it is done for the `arch` flavor of Linux).

However, that comes with its own set of problems as now also, say,
`msys2-runtime` would be upgraded, which frequently causes issues
because MSYS2 actually has a two-stage upgrade process, where only
"system" packages (such as "msys2-runtime") are upgraded in the first
stage, and the rest of the packages (e.g. "tmate") is installed only in
the second stage, which means that `pacman` would have to be invoked
_twice_.

Instead, let's just tell Pacman to stop looking for the latest packages
by skipping the `-y` flag that tells it to first to upgrade the package
index. That way, we get the "tmate" version that goes with the currently
installed "libopenssl" version, even if neither are the very latest
versions but potentially a week or three behind.