Support (and default to) limit-access-to-actor == auto
It is relatively easy to forget to limit access to the tmate session and
accidentally make secrets accessible to anyone with an SSH client.

Let's introduce a `limit-access-to-actor` mode that defaults to using
SSH public keys registered with the actor's GitHub profile, if any, and
fall back to using potentially unsafe tmate session that anybdy can
connect to.

This strikes a balance between security and ease of use.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>