When using public key authorization, show also `-i` cmdline
The line shown by `action-tmate` with the `ssh` invocation assumes that
either no public key authorization was enabled, or that there is only
one private key present locally that `ssh` picks up by default.

However, many developers use multiple, distinct private keys for
distinct purposes, so that one compromised private key does not make
_all_ of the hosts vulnerable that are accessible to the user.

In such an instance, the private key must be specified explicitly via
the `-i` option.

The message shown by `action-tmate` is meant to be copy/paste'able, but
for invocations requiring the `-i` option that is not true. We cannot
_really_ make it completely copy/paste'able, but we can at least make it
easier to copy/paste/edit the invocation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>