Warn more strongly when no public SSH key was found
It is very, very easy to start an action-tmate session and forgetting
that secrets are made available that way that should be protected.

For example, `actions/checkout` persists credentials by default, and if
the tmate session is not secured, theoretically anybody could exfiltrate
those credentials.

Let's warn a lot louder about this and strongly encourage allowing
authenticated access only.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>