Start tmate with the '-F' option as suggested in the docs
See <https://tmate.io/#remote_access>. This option seems to be
made for this exact use-case. This commit is merely a suggestion
that the -F option could be used, though. If it were to be used,
the implementation around the line changed in this commit would
still need to be adjusted, of course.