Implement a "detached" mode
In "detached" mode, the `tmate` session will be created, the Action will
then print the information how to connect, and then exit with success.

The workflow job's steps will run concurrently with the `tmate` session,
allowing the user to monitor the steps as they are progressing.

At the end of the job, the Action's newly-introduced post-job Action
will wait until the session is ended, or up to 10 minutes for any client
to connect at all.

This detached mode is inspired by CircleCI/CirrusCI's debugging support
that _only_ allows a concurrent shell terminal to be opened.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>